First stab at non-singleton tree.

+++ src/interaction_specific/log2d/module_interaction_specific.f90
@@ -283,8 +283,8 @@ module module_interaction_specific
       !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
       subroutine get_number_of_interactions_per_particle(npart_total, nintmax)
         implicit none
-        integer, intent(in) :: npart_total !< total number of particles
-        integer, intent(out) :: nintmax !< maximum number of interactions per particle
+        integer*8, intent(in) :: npart_total !< total number of particles
+        integer*8, intent(out) :: nintmax !< maximum number of interactions per particle
 
         real*8 :: invnintmax !< inverse of nintmax to avoid division by zero for theta == 0.0
